+++ narrenschiff/modules/helm.py
@@ -1,6 +1,9 @@
+import os
 import subprocess
+from contextlib import suppress
 
 from narrenschiff.modules.common import NarrenschiffModule
+from narrenschiff.secretmap import SecretmapCommand
 
 
 class Helm(NarrenschiffModule):
@@ -13,6 +16,8 @@ class Helm(NarrenschiffModule):
         chart = self.command.get('chart', '')
         options = self.command.get('opts')
         arguments = self.command.get('args', {})
+
+        self.parse_and_copy_secretmaps_args()
 
         try:
             args_set = self.command.get('args').get('set')
@@ -35,3 +40,20 @@ class Helm(NarrenschiffModule):
 
         cmd = ' '.join([Helm.helm_cmd, command, chart, opts, args, sets])
         subprocess.run(cmd, shell=True, check=True)
+
+    def parse_and_copy_secretmaps_args(self):
+        """
+        Mutate secretmap arguments. Expand secretmap paths to match files in
+        the ``/tmp`` directory.
+
+        :return: Void
+        :rtype: ``None``
+        """
+        secretmap = '{{secretmap}}/'
+        for key, value in self.command.get('args', {}).items():
+            with suppress(AttributeError):
+                if value.startswith(secretmap):
+                    basepath = value.replace(secretmap, '')
+                    tmp = SecretmapCommand().tmp
+                    rendered_path = os.path.join(tmp, basepath)
+                    self.command['args'][key] = rendered_path