+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 0.1.0 - 27.8.2019.
+
+## Added
+
+* Secretmap now decrypts and copies files to `/tmp`
+
+## Fixed
+
+* Secretmap file is now storing _relative_ paths
+
 # 0.0.5 - 27.8.2019.
 
 ## Fixed